create like and follow features

+++ server/src/services/favorite.service.js
@@ -3,14 +3,14 @@ import { Favorite } from '../models/index.js';
 
 const favoriteService = {
   getAll: async (params = {}) => {
-    const { _limit, _sort, _order } = params;
+    const { _page, _limit, _sort, _order, _fields, ...others } = params;
 
     if (_limit || (_sort && _order)) {
       const response = await paginateSort(params, Favorite);
       return response;
     }
 
-    const response = await Favorite.find(params);
+    const response = await Favorite.find(others).select(_fields);
     return { data: response };
   },
   add: async (userId = '', chapterId = '') => {
@@ -23,7 +23,7 @@ const favoriteService = {
     return response;
   },
   delete: async (userId, chapterId) => {
-    const response = await Favorite.findOneAndDelete({ userId, chapterId });
+    const response = await Favorite.findOneAndDelete({ user_id: userId, chapter_id: chapterId });
     return response;
   },
 };
